retain only single mcp-discord-agent using sampling
